+++ aiomygas/queries/accounts.py
@@ -1,4 +1,6 @@
 """GraphQL queries for accounts."""
+from __future__ import annotations
+
 from .base import BaseQuery