+++ aiomygas/queries/charges.py
@@ -1,4 +1,6 @@
 """GraphQL queries for charges."""
+from __future__ import annotations
+
 from .base import BaseQuery
 from ..const import ATTR_LSPU_ID